+++ package.json
@@ -2,12 +2,12 @@
   "name": "mercosur-plates",
   "version": "1.0.0",
   "description": "Interactive visualization of Brazilian Portuguese words which fit in Mercosur's registration plates.",
-  "main": "index.js",
+  "main": "./js/main.js",
   "directories": {
     "doc": "docs"
   },
   "scripts": {
-    "start": "node index.js",
+    "start": "node ./js/main.js",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "repository": {
